+++ asv/benchmarks.py
@@ -192,12 +192,16 @@ class Benchmarks(dict):
                 try:
                     env.install_project(conf, repo, commit_hash)
 
+                    env_vars = dict(os.environ)
+                    env_vars.update(env.env_vars)
+
                     result_file = os.path.join(result_dir, 'result.json')
                     env.run(
                         [runner.BENCHMARK_RUN_SCRIPT, 'discover',
                          os.path.abspath(root),
                          os.path.abspath(result_file)],
                         cwd=result_dir,
+                        env=env_vars,
                         dots=False)
 
                     try:
@@ -228,6 +232,7 @@ class Benchmarks(dict):
                          os.path.abspath(root)],
                         cwd=result_dir,
                         dots=False,
+                        env=env_vars,
                         valid_return_codes=None,
                         return_stderr=True,
                         redirect_stderr=True)