+++ asv/runner.py
@@ -667,6 +667,9 @@ class Spawner(object):
     def create_setup_cache(self, benchmark_id, timeout):
         cache_dir = tempfile.mkdtemp()
 
+        env_vars = dict(os.environ)
+        env_vars.update(self.env.env_vars)
+
         out, _, errcode = self.env.run(
             [BENCHMARK_RUN_SCRIPT, 'setup_cache',
              os.path.abspath(self.benchmark_dir),
@@ -674,7 +677,9 @@ class Spawner(object):
             dots=False, display_error=False,
             return_stderr=True, valid_return_codes=None,
             redirect_stderr=True,
-            cwd=cache_dir, timeout=timeout)
+            cwd=cache_dir,
+            timeout=timeout,
+            env=env_vars)
 
         if errcode == 0:
             return cache_dir, None
@@ -683,12 +688,16 @@ class Spawner(object):
             return None, out.strip()
 
     def run(self, name, params_str, profile_path, result_file_name, timeout, cwd):
+        env_vars = dict(os.environ)
+        env_vars.update(self.env.env_vars)
+
         out, _, errcode = self.env.run(
             [BENCHMARK_RUN_SCRIPT, 'run', os.path.abspath(self.benchmark_dir),
              name, params_str, profile_path, result_file_name],
             dots=False, timeout=timeout,
             display_error=False, return_stderr=True, redirect_stderr=True,
-            valid_return_codes=None, cwd=cwd)
+            valid_return_codes=None, cwd=cwd,
+            env=env_vars)
         return out, errcode
 
     def preimport(self):
@@ -708,9 +717,14 @@ class ForkServer(Spawner):
         self.tmp_dir = tempfile.mkdtemp(prefix='asv-forkserver-')
         self.socket_name = os.path.join(self.tmp_dir, 'socket')
 
+        env_vars = dict(os.environ)
+        env_vars.update(env.env_vars)
+
         self.server_proc = env.run(
             [BENCHMARK_RUN_SCRIPT, 'run_server', self.benchmark_dir, self.socket_name],
-            return_popen=True, redirect_stderr=True)
+            return_popen=True,
+            redirect_stderr=True,
+            env=env_vars)
 
         self._server_output = None
         self.stdout_reader_thread = threading.Thread(target=self._stdout_reader)