+++ test/test_environment.py
@@ -9,6 +9,7 @@ import sys
 import six
 import pytest
 import json
+from collections import defaultdict
 
 from asv import config
 from asv import environment
@@ -143,13 +144,17 @@ def test_matrix_expand_basic():
         'pkg5': []
     }
 
-    combinations = _sorted_dict_list(environment.iter_requirement_matrix(
+    combinations = _sorted_dict_list(environment.iter_matrix(
         conf.environment_type, conf.pythons, conf))
     expected = _sorted_dict_list([
-        {'python': '2.6', 'pkg2': '', 'pkg3': '', 'pkg4': '1.2', 'pkg5': ''},
-        {'python': '2.6', 'pkg2': '', 'pkg3': '', 'pkg4': '3.4', 'pkg5': ''},
-        {'python': '2.7', 'pkg2': '', 'pkg3': '', 'pkg4': '1.2', 'pkg5': ''},
-        {'python': '2.7', 'pkg2': '', 'pkg3': '', 'pkg4': '3.4', 'pkg5': ''},
+        {('python', None): '2.6', ('req', 'pkg2'): '', ('req', 'pkg3'): '',
+         ('req', 'pkg4'): '1.2', ('req', 'pkg5'): ''},
+        {('python', None): '2.6', ('req', 'pkg2'): '', ('req', 'pkg3'): '',
+         ('req', 'pkg4'): '3.4', ('req', 'pkg5'): ''},
+        {('python', None): '2.7', ('req', 'pkg2'): '', ('req', 'pkg3'): '',
+         ('req', 'pkg4'): '1.2', ('req', 'pkg5'): ''},
+        {('python', None): '2.7', ('req', 'pkg2'): '', ('req', 'pkg3'): '',
+         ('req', 'pkg4'): '3.4', ('req', 'pkg5'): ''},
     ])
     assert combinations == expected
 
@@ -167,13 +172,13 @@ def test_matrix_expand_include():
         {'environment_type': 'something', 'python': '2.7', 'b': '5'},
     ]
 
-    combinations = _sorted_dict_list(environment.iter_requirement_matrix(
+    combinations = _sorted_dict_list(environment.iter_matrix(
         conf.environment_type, conf.pythons, conf))
     expected = _sorted_dict_list([
-        {'python': '2.6', 'a': '1'},
-        {'python': '3.5', 'b': '2'},
-        {'python': '2.7', 'b': '3'},
-        {'python': '2.7', 'b': '5'}
+        {('python', None): '2.6', ('req', 'a'): '1'},
+        {('python', None): '3.5', ('req', 'b'): '2'},
+        {('python', None): '2.7', ('req', 'b'): '3'},
+        {('python', None): '2.7', ('req', 'b'): '5'}
     ])
     assert combinations == expected
 
@@ -181,7 +186,7 @@ def test_matrix_expand_include():
         {'b': '2'}
     ]
     with pytest.raises(util.UserError):
-        list(environment.iter_requirement_matrix(conf.environment_type, conf.pythons, conf))
+        list(environment.iter_matrix(conf.environment_type, conf.pythons, conf))
 
 
 def test_matrix_expand_include_detect_env_type():
@@ -194,10 +199,10 @@ def test_matrix_expand_include_detect_env_type():
         {'sys_platform': sys.platform, 'python': PYTHON_VER1},
     ]
 
-    combinations = _sorted_dict_list(environment.iter_requirement_matrix(
+    combinations = _sorted_dict_list(environment.iter_matrix(
         conf.environment_type, conf.pythons, conf))
     expected = _sorted_dict_list([
-        {'python': PYTHON_VER1},
+        {('python', None): PYTHON_VER1},
     ])
     assert combinations == expected
 
@@ -220,11 +225,11 @@ def test_matrix_expand_exclude():
         {'python': '2.7', 'b': None},
         {'python': '2.6', 'a': '1'},
     ]
-    combinations = _sorted_dict_list(environment.iter_requirement_matrix(
+    combinations = _sorted_dict_list(environment.iter_matrix(
         conf.environment_type, conf.pythons, conf))
     expected = _sorted_dict_list([
-        {'python': '2.7', 'a': '1', 'b': '1'},
-        {'python': '2.7', 'b': '2'}
+        {('python', None): '2.7', ('req', 'a'): '1', ('req', 'b'): '1'},
+        {('python', None): '2.7', ('req', 'b'): '2'}
     ])
     assert combinations == expected
 
@@ -232,12 +237,12 @@ def test_matrix_expand_exclude():
     conf.exclude = [
         {'python': '.*', 'b': None},
     ]
-    combinations = _sorted_dict_list(environment.iter_requirement_matrix(
+    combinations = _sorted_dict_list(environment.iter_matrix(
         conf.environment_type, conf.pythons, conf))
     expected = _sorted_dict_list([
-        {'python': '2.6', 'a': '1', 'b': '1'},
-        {'python': '2.7', 'a': '1', 'b': '1'},
-        {'python': '2.7', 'b': '2'}
+        {('python', None): '2.6', ('req', 'a'): '1', ('req', 'b'): '1'},
+        {('python', None): '2.7', ('req', 'a'): '1', ('req', 'b'): '1'},
+        {('python', None): '2.7', ('req', 'b'): '2'}
     ])
     assert combinations == expected
 
@@ -245,10 +250,10 @@ def test_matrix_expand_exclude():
     conf.exclude = [
         {'environment_type': 'some.*'},
     ]
-    combinations = _sorted_dict_list(environment.iter_requirement_matrix(
+    combinations = _sorted_dict_list(environment.iter_matrix(
         conf.environment_type, conf.pythons, conf))
     expected = [
-        {'python': '2.7', 'b': '2'}
+        {('python', None): '2.7', ('req', 'b'): '2'}
     ]
     assert combinations == expected
 
@@ -256,10 +261,10 @@ def test_matrix_expand_exclude():
     conf.exclude = [
         {'sys_platform': sys.platform},
     ]
-    combinations = _sorted_dict_list(environment.iter_requirement_matrix(
+    combinations = _sorted_dict_list(environment.iter_matrix(
         conf.environment_type, conf.pythons, conf))
     expected = [
-        {'python': '2.7', 'b': '2'}
+        {('python', None): '2.7', ('req', 'b'): '2'}
     ]
     assert combinations == expected
 
@@ -267,14 +272,46 @@ def test_matrix_expand_exclude():
     conf.exclude = [
         {'python': '(?!2.6).*'}
     ]
-    combinations = _sorted_dict_list(environment.iter_requirement_matrix(
+    combinations = _sorted_dict_list(environment.iter_matrix(
         conf.environment_type, conf.pythons, conf))
     expected = _sorted_dict_list([
-        {'python': '2.6', 'a': '1', 'b': '1'},
-        {'python': '2.6', 'a': '1'},
-        {'python': '2.7', 'b': '2'}
+        {('python', None): '2.6', ('req', 'a'): '1', ('req', 'b'): '1'},
+        {('python', None): '2.6', ('req', 'a'): '1'},
+        {('python', None): '2.7', ('req', 'b'): '2'}
     ])
     assert combinations == expected
+
+
+def test_iter_env_matrix_combinations():
+    conf = config.Config()
+    conf.environment_type = 'something'
+    conf.pythons = ["2.6"]
+    conf.matrix = {}
+    conf.include = []
+
+    # (matrix, expected)
+    env_matrices = [
+        ({'var0': ['val0', 'val1'], 'var1': ['val2', 'val3']},
+         [{'var0': 'val0', 'var1': 'val2'},
+          {'var0': 'val0', 'var1': 'val3'},
+          {'var0': 'val1', 'var1': 'val2'},
+          {'var0': 'val1', 'var1': 'val3'}]),
+        ({'var0': ['val0', 'val1'], 'var1': ['val2', None]},
+         [{'var0': 'val0', 'var1': 'val2'}, {'var0': 'val0'},
+          {'var0': 'val1', 'var1': 'val2'}, {'var0': 'val1'}]),
+        ({'var0': ['val0', 'val1']},
+         [{'var0': 'val0'}, {'var0': 'val1'}]),
+        ({}, [{}]),
+    ]
+
+    for matrix, expected in env_matrices:
+        conf.matrix = {'@env': matrix}
+        expected = [{('env', key): value for key, value in item.items()}
+                    for item in expected]
+        for m in expected:
+            m['python', None] = "2.6"
+        result = _sorted_dict_list(environment.iter_matrix(conf.environment_type, conf.pythons, conf))
+        assert result == _sorted_dict_list(expected)
 
 
 @pytest.mark.skipif((not HAS_CONDA), reason="Requires conda")
@@ -473,6 +510,7 @@ def test_matrix_existing():
     environments = list(environment.get_environments(conf, None))
     items = [(env.tool_name, tuple(env.requirements.keys())) for env in environments]
     assert items == [('existing', ())]
+
 
 # environment.yml should respect the specified order
 # of channels when adding packages
@@ -736,14 +774,14 @@ def test_installed_commit_hash(tmpdir):
 
     # Check updating installed_commit_hash
     assert env.installed_commit_hash == None
-    assert env._env_vars.get('ASV_COMMIT') == None
+    assert env._global_env_vars.get('ASV_COMMIT') == None
     env.install_project(conf, repo, commit_hash)
     assert env.installed_commit_hash == commit_hash
-    assert env._env_vars.get('ASV_COMMIT') == commit_hash
+    assert env._global_env_vars.get('ASV_COMMIT') == commit_hash
 
     env = get_env()
     assert env.installed_commit_hash == commit_hash
-    assert env._env_vars.get('ASV_COMMIT') == commit_hash
+    assert env._global_env_vars.get('ASV_COMMIT') == commit_hash
 
     # Configuration change results to reinstall
     env._project = "something"
@@ -753,11 +791,11 @@ def test_installed_commit_hash(tmpdir):
     env = get_env()
     env._uninstall_project()
     assert env.installed_commit_hash == None
-    assert env._env_vars.get('ASV_COMMIT') != None
+    assert env._global_env_vars.get('ASV_COMMIT') != None
 
     env = get_env()
     assert env.installed_commit_hash == None
-    assert env._env_vars.get('ASV_COMMIT') == None
+    assert env._global_env_vars.get('ASV_COMMIT') == None
 
 
 def test_install_success(tmpdir):
@@ -783,3 +821,77 @@ def test_install_success(tmpdir):
     env.install_project(conf, repo, commit_hash)
 
     env.run(['-c', 'import asv_test_repo as t, sys; sys.exit(0 if t.dummy_value == 0 else 1)'])
+
+
+def test_install_env_matrix_values(tmpdir):
+    tmpdir = six.text_type(tmpdir)
+
+    dvcs = generate_test_repo(tmpdir, [0], dvcs_type='git')
+    commit_hash = dvcs.get_branch_hashes()[0]
+
+    conf = config.Config()
+    conf.env_dir = os.path.join(tmpdir, "env")
+    conf.pythons = [PYTHON_VER1]
+    conf.repo = os.path.abspath(dvcs.path)
+    conf.matrix = {'@env': {'SOME_ASV_TEST_BUILD_VALUE': '1'},
+                   '@env_nobuild': {'SOME_ASV_TEST_NON_BUILD_VALUE': '1'}}
+
+    repo = get_repo(conf)
+
+    env = list(environment.get_environments(conf, None))[0]
+    env.create()
+    env.install_project(conf, repo, commit_hash)
+
+    env.run(['-c',
+             'import asv_test_repo.build_time_env as t, sys; '
+             'sys.exit(0 if t.env["SOME_ASV_TEST_BUILD_VALUE"] == "1" else 1)'])
+
+    env.run(['-c',
+             'import asv_test_repo.build_time_env as t, sys; '
+             'sys.exit(0 if "SOME_ASV_TEST_NON_BUILD_VALUE" not in t.env else 1)'])
+
+
+def test_environment_env_matrix():
+    # (build_vars, non_build_vars, environ_count, build_count)
+    configs = [
+        ({}, {}, 1, 1),
+        ({"var1": ["val1"]}, {}, 1, 1),
+        ({"var1": ["val1", "val2", "val3"]}, {}, 3, 3),
+        ({"var1": ["val1", "val2"], "var2": ['val3', 'val4']}, {}, 4, 4),
+        ({"var1": ["val1", "val2"], "var2": ['val3', None]}, {}, 4, 4),
+        ({"var1": ["val1", "val2"]}, {"var2": ['val3', None]}, 4, 2),
+        ({"var1": ["val1", "val2"], "var2": ['val3', 'val4']},
+         {"var3": ['val5', None]}, 8, 4),
+    ]
+
+    for build_vars, non_build_vars, environ_count, build_count in configs:
+        conf = config.Config()
+
+        conf.matrix = {
+            "@env": build_vars,
+            "@env_nobuild": non_build_vars,
+        }
+        environments = list(environment.get_environments(conf, None))
+
+        assert len(environments) == environ_count
+        assert len(set(e.dir_name for e in environments)) == build_count
+
+
+def test__parse_exclude_include_rule():
+    cases = [
+        ({"python": "2.6", "environment_type": "conda", "sys_platform": "123",
+          "@env": {"A": "B"}, "@env_nobuild": {"C": "D"}, "foo": "9"},
+         {("python", None): "2.6", ("environment_type", None): "conda", ("sys_platform", None): "123",
+          ("env", "A"): "B", ("env_nobuild", "C"): "D", ("req", "foo"): "9"})
+    ]
+    for rule, expected in cases:
+        parsed = environment._parse_exclude_include_rule(rule)
+        assert parsed == expected
+
+
+def test__parse_matrix_entries():
+    entries =  {("python", None): "2.6",  ("env", "A"): "B", ("env_nobuild", "C"): "D", ("req", "foo"): "9"}
+    python, requirements, tagged_env_vars = environment._parse_matrix_entries(entries)
+    assert python == "2.6"
+    assert requirements == {"foo": "9"}
+    assert tagged_env_vars == {("build", "A"): "B", ("nobuild", "C"): "D"}