+++ test/test_run.py
@@ -272,7 +272,6 @@ def test_cpu_affinity(basic_conf):
     # Only one environment
     conf.matrix = {}
 
-    # Tests multiple calls to "asv run --append-samples"
     tools.run_asv_with_conf(conf, 'run', "master^!",
                             '--bench', 'time_examples.TimeSuite.time_example_benchmark_1',
                             '--cpu-affinity=0', '-a', 'repeat=(1, 1, 10.0)', '-a', 'processes=1',
@@ -286,3 +285,42 @@ def test_cpu_affinity(basic_conf):
                   if fn != 'machine.json']
     data = util.load_json(result_fn)
     assert data['results']['time_examples.TimeSuite.time_example_benchmark_1']
+
+
+def test_env_matrix_value(basic_conf):
+    tmpdir, local, conf, machine_file = basic_conf
+
+    conf.matrix = {}
+
+    def check_env_matrix(env_build, env_nobuild):
+        conf.matrix = {"@env": env_build, "@env_nobuild": env_nobuild}
+
+        tools.run_asv_with_conf(conf, 'run', "master^!",
+                                '--bench', 'time_secondary.track_environment_value',
+                                _machine_file=machine_file)
+
+        # Check run produced a result
+        result_dir = join(tmpdir, 'results_workflow', 'orangutan')
+
+        result_fn1, = glob.glob(result_dir + '/*-SOME_TEST_VAR1.json')
+        result_fn2, = glob.glob(result_dir + '/*-SOME_TEST_VAR2.json')
+
+        data = util.load_json(result_fn1)
+        assert data['results']['time_secondary.track_environment_value'] == 1
+
+        data = util.load_json(result_fn2)
+        assert data['results']['time_secondary.track_environment_value'] == 2
+
+    check_env_matrix({}, {'SOME_TEST_VAR': ['1', '2']})
+    check_env_matrix({'SOME_TEST_VAR': ['1', '2']}, {})
+
+
+def test_parallel(basic_conf, dummy_packages):
+    tmpdir, local, conf, machine_file = basic_conf
+
+    conf.matrix["@env"] = {"SOME_TEST_VAR": ["1", "2"]}
+    conf.matrix["@env_nobuild"] = {"SOME_OTHER_TEST_VAR": ["1", "2"]}
+
+    tools.run_asv_with_conf(conf, 'run', "master^!",
+                            '--bench', 'time_secondary.track_environment_value',
+                            '--parallel=2', _machine_file=machine_file)